refactor: move Sequelize query executor out of DAL into backend
