+++ backend/src/api/member/memberMerge.ts
@@ -1,5 +1,6 @@
 import { CommonMemberService, invalidateMemberQueryCache } from '@crowd/common_services'
-import { optionsQx } from '@crowd/data-access-layer'
+
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 
 import Permissions from '../../security/permissions'
 import track from '../../segment/track'
