+++ backend/src/api/public/v1/affiliations/getAffiliationByHandle.ts
@@ -4,10 +4,10 @@ import { NotFoundError } from '@crowd/common'
 import {
   findMembersByGithubHandles,
   findVerifiedEmailsByMemberIds,
-  optionsQx,
   resolveAffiliationsByMemberIds,
 } from '@crowd/data-access-layer'
 
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 import { ok } from '@/utils/api'
 
 export async function getAffiliationByHandle(req: Request, res: Response): Promise<void> {
