+++ backend/src/api/public/v1/members/project-affiliations/patchProjectAffiliation.ts
@@ -11,11 +11,11 @@ import {
   findMaintainerRoles,
   findMemberById,
   insertMemberSegmentAffiliations,
-  optionsQx,
 } from '@crowd/data-access-layer'
 import type { ISegmentAffiliationWithOrg } from '@crowd/data-access-layer'
 import { deleteMemberSegmentAffiliations } from '@crowd/data-access-layer/src/member_segment_affiliations'
 
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 import { ok } from '@/utils/api'
 import { validateOrThrow } from '@/utils/validation'
