+++ backend/src/api/public/v1/members/resolveMember.ts
@@ -2,9 +2,10 @@ import type { Request, Response } from 'express'
 import { z } from 'zod'
 
 import { ConflictError, NotFoundError } from '@crowd/common'
-import { findMemberIdsByIdentities, optionsQx } from '@crowd/data-access-layer'
+import { findMemberIdsByIdentities } from '@crowd/data-access-layer'
 import { IMemberIdentity, MemberIdentityType, PlatformType } from '@crowd/types'
 
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 import { ok } from '@/utils/api'
 import { validateOrThrow } from '@/utils/validation'
