+++ backend/src/api/public/v1/members/work-experiences/getMemberWorkExperiences.ts
@@ -6,9 +6,9 @@ import {
   MemberField,
   fetchManyMemberOrgsWithOrgData,
   findMemberById,
-  optionsQx,
 } from '@crowd/data-access-layer'
 
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 import { ok } from '@/utils/api'
 import { groupMemberOrganizations, toMemberWorkExperience } from '@/utils/mapper'
 import { validateOrThrow } from '@/utils/validation'
