+++ backend/src/api/public/v1/members/work-experiences/verifyMemberWorkExperience.ts
@@ -10,11 +10,11 @@ import {
   fetchManyMemberOrgsWithOrgData,
   fetchMemberOrganizations,
   findMemberById,
-  optionsQx,
   updateMemberOrganization,
 } from '@crowd/data-access-layer'
 import { IMemberOrganization, IMemberRoleWithOrganization } from '@crowd/types'
 
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 import { ok } from '@/utils/api'
 import {
   getOverlappingGroupedMemberOrganizations,
