+++ backend/src/api/public/v1/organizations/getOrganization.ts
@@ -2,8 +2,9 @@ import type { Request, Response } from 'express'
 import { z } from 'zod'
 
 import { BadRequestError, NotFoundError, normalizeHostname } from '@crowd/common'
-import { findOrganizationByNameOrDomain, optionsQx } from '@crowd/data-access-layer'
+import { findOrganizationByNameOrDomain } from '@crowd/data-access-layer'
 
+import { optionsQx } from '@/database/sequelizeQueryExecutor'
 import { ok } from '@/utils/api'
 import { validateOrThrow } from '@/utils/validation'
