+++ backend/src/database/repositories/repositoryBase.ts
@@ -1,13 +1,10 @@
 /* eslint-disable class-methods-use-this,@typescript-eslint/no-unused-vars */
 import { Sequelize } from 'sequelize'
 
-import {
-  QueryExecutor,
-  SequelizeQueryExecutor,
-  TransactionalSequelizeQueryExecutor,
-  optionsQx,
-} from '@crowd/data-access-layer/src/queryExecutor'
+import { QueryExecutor } from '@crowd/data-access-layer/src/queryExecutor'
 import { PageData, SearchCriteria } from '@crowd/types'
+
+import { optionsQx } from '../sequelizeQueryExecutor'
 
 import { IRepositoryOptions } from './IRepositoryOptions'
 import SequelizeRepository from './sequelizeRepository'
