+++ services/libs/data-access-layer/package.json
@@ -34,7 +34,6 @@
   },
   "devDependencies": {
     "@types/node": "^20.8.2",
-    "sequelize": "6.37.8",
     "typescript": "^5.6.3"
   }
 }
